Limit todos (#15)
* Add Todo Filtering by Owner w/ Tests
Close #5
Owner filtering is now implemented for todos and
the tests all pass.

* Add Todo Filtering by Owner w/ Tests

Owner filtering is now implemented for todos and tests for
this feature have been written and pass.

* Implement limitation methods with testing

Limiting todos is showing functionality as demonstrated
by the passing tests written for it. The first test limits
all todos by calling the limiting helper function
and the second places the limit into the query parameters array to be
parsed by the listTodo method.